universal_payload.rst: Add ClockRate field to serial port info
Add a ClockRate field to the UNIVERSAL_PAYLOAD_SERIAL_PORT_INFO struct
to allow the host firmware to pass the serial clock rate to the payload
in order to properly initialize the serial/debug port.

This field is necessary because AMD Zen platform devices do not use
the same clock rate as Intel devices.

Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>